+++ CHANGELOG.md
@@ -1,3 +1,20 @@
+## 1.43.1
+
+* No user-visible changes.
+
+## 1.43.0
+
+### JS API
+
+* Add support for the `logger` option. This takes an object that can define
+  `warn` or `debug` methods to add custom handling for messages emitted by the
+  Sass compiler. See [the JS API docs] for details.
+
+  [the JS API docs]: https://sass-lang.com/documentation/js-api/interfaces/Logger
+
+* Add a `Logger.silent` object that can be passed to the `logger` option to
+  silence all messages from the Sass compiler.
+
 ## 1.42.1
 
 * Fix a bug where Sass variables and function calls in calculations weren't
@@ -98,7 +115,7 @@
 
 * Partial fix for a bug where `@at-root` does not work properly in nested
   imports that contain `@use` rules. If the only `@use` rules in the nested
-  import are for built-in modules, `@at-root` should now work properly. 
+  import are for built-in modules, `@at-root` should now work properly.
 
 ## 1.39.0